+++ clang-tidy/nodecpp/NewExprCheck.cpp
@@ -1,4 +1,4 @@
-//===--- NewNakedPtrCheck.cpp - clang-tidy---------------------------------===//
+//===--- NewExprCheck.cpp - clang-tidy-------------------------------------===//
 //
 //                     The LLVM Compiler Infrastructure
 //
@@ -7,7 +7,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "NewNakedPtrCheck.h"
+#include "NewExprCheck.h"
 #include "clang/AST/ASTContext.h"
 #include "clang/ASTMatchers/ASTMatchFinder.h"
 
@@ -17,7 +17,7 @@ namespace clang {
 namespace tidy {
 namespace nodecpp {
 
-void NewNakedPtrCheck::registerMatchers(MatchFinder *Finder) {
+void NewExprCheck::registerMatchers(MatchFinder *Finder) {
 
   if (!getLangOpts().CPlusPlus)
     return;
@@ -34,14 +34,14 @@ void NewNakedPtrCheck::registerMatchers(MatchFinder *Finder) {
 
   const auto badNew = cxxNewExpr(
 	  unless(anyOf(hasParent(OwnCtor),
-		  hasParent(OwnReset)))).bind("badNew");
+		  hasParent(OwnReset)))).bind("expr");
 
   Finder->addMatcher(badNew, this);
 }
 
-void NewNakedPtrCheck::check(const MatchFinder::MatchResult &Result) {
+void NewExprCheck::check(const MatchFinder::MatchResult &Result) {
  
-  const auto *MatchedDecl = Result.Nodes.getNodeAs<CXXNewExpr>("badNew");
+  const auto *MatchedDecl = Result.Nodes.getNodeAs<CXXNewExpr>("expr");
 
   diag(MatchedDecl->getLocStart(), "new expression must be owned by unique_ptr");
 }